+++ flag/service/slack/slack.go
@@ -2,4 +2,5 @@ package slack
 
 type Slack struct {
 	ApiToken string
+	ApiUrl   string
 }