+++ service/controller/managementcluster/resource.go
@@ -61,6 +61,7 @@ type resourcesConfig struct {
 	GrafanaAddress string
 	OpsgenieKey    string
 	SlackApiToken  string
+	SlackApiUrl    string
 
 	MimirEnabled bool
 
@@ -140,6 +141,7 @@ func newResources(config resourcesConfig) ([]resource.Interface, error) {
 			Pipeline:       config.Pipeline,
 			Proxy:          config.Proxy,
 			SlackApiToken:  config.SlackApiToken,
+			SlackApiUrl:    config.SlackApiUrl,
 		}
 
 		alertmanagerConfigResource, err = alertmanagerconfig.New(c)