+++ service/service.go
@@ -216,6 +216,7 @@ func New(config Config) (*Service, error) {
 			GrafanaAddress: config.Viper.GetString(config.Flag.Service.Grafana.Address),
 			OpsgenieKey:    config.Viper.GetString(config.Flag.Service.Opsgenie.Key),
 			SlackApiToken:  config.Viper.GetString(config.Flag.Service.Slack.ApiToken),
+			SlackApiUrl:    config.Viper.GetString(config.Flag.Service.Slack.ApiUrl),
 
 			MimirEnabled: config.Viper.GetBool(config.Flag.Service.Mimir.Enabled),