+++ flag/service/slack/slack.go
@@ -1,5 +1,6 @@
 package slack
 
 type Slack struct {
-	ApiURL string
+	ApiToken string
+	ApiURL   string
 }