+++ service/controller/resource/alerting/alertmanagerconfig/resource_test.go
@@ -180,3 +180,49 @@ func TestRenderingOfAlertmanagerConfigWithMimirEnabled(t *testing.T) {
 		}
 	}
 }
+
+func TestRenderingOfAlertmanagerConfigSlackToken(t *testing.T) {
+	var testFunc unittest.TestFunc
+	{
+
+		proxyConfig := httpproxy.Config{}
+
+		config := Config{
+			GrafanaAddress: "https://grafana",
+			Installation:   "test-installation",
+			OpsgenieKey:    "opsgenie-key",
+			Proxy:          proxyConfig.ProxyFunc(),
+			Pipeline:       "testing",
+			SlackApiURL:    "https://slack",
+			SlackApiToken:  "some-token",
+		}
+		testFunc = func(v interface{}) (interface{}, error) {
+			return renderAlertmanagerConfig(unittest.ProjectRoot(), config)
+		}
+	}
+
+	for _, flavor := range unittest.ProviderFlavors {
+		outputDir, err := filepath.Abs("./test/alertmanager-config/slack-token/" + flavor)
+		if err != nil {
+			t.Fatal(err)
+		}
+
+		c := unittest.Config{
+			OutputDir:            outputDir,
+			T:                    t,
+			TestFunc:             testFunc,
+			TestFuncReturnsBytes: true,
+			Flavor:               flavor,
+			Update:               *update,
+		}
+		runner, err := unittest.NewRunner(c)
+		if err != nil {
+			t.Fatal(err)
+		}
+
+		err = runner.Run()
+		if err != nil {
+			t.Fatal(err)
+		}
+	}
+}