+++ service/controller/managementcluster/controller.go
@@ -18,6 +18,7 @@ import (
 
 	"github.com/giantswarm/prometheus-meta-operator/v2/pkg/cluster"
 	"github.com/giantswarm/prometheus-meta-operator/v2/pkg/project"
+	"github.com/giantswarm/prometheus-meta-operator/v2/service/controller/resource/monitoring/remotewriteconfig"
 	"github.com/giantswarm/prometheus-meta-operator/v2/service/key"
 )
 
@@ -53,6 +54,8 @@ type ControllerConfig struct {
 	PrometheusScrapeInterval     string
 	PrometheusImageRepository    string
 	PrometheusVersion            string
+
+	PrometheusAgentShardingStrategy remotewriteconfig.PrometheusAgentShardingStrategy
 
 	RestrictedAccessEnabled bool
 	WhitelistedSubnets      string